Added ability to search in files
